crud commentaires

+++ titrePro/isoPhoto/back/src/main/java/co/simplon/titrepro/isophoto/api/controller/CommentaireController.java
@@ -2,11 +2,10 @@ package co.simplon.titrepro.isophoto.api.controller;
 
 import java.util.List;
 
-import javax.validation.Valid;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,14 +15,23 @@ import org.springframework.web.bind.annotation.RestController;
 
 import co.simplon.titrepro.isophoto.model.Commentaire;
 import co.simplon.titrepro.isophoto.repository.CommentaireRepository;
+import co.simplon.titrepro.isophoto.repository.PhotoRepository;
+import co.simplon.titrepro.isophoto.repository.PhotographeRepository;
 import co.simplon.titrepro.isophoto.services.CommentaireService;
 
 @RestController
+@CrossOrigin("http://localhost:4200")
 @RequestMapping("/api")
 public class CommentaireController {
 
 	@Autowired
 	CommentaireRepository commentaireRepo;
+	
+	@Autowired
+	PhotoRepository photoRepo;
+	
+	@Autowired
+	PhotographeRepository photographeRepo;
 
 	private CommentaireService commentaireService;
 
@@ -43,33 +51,49 @@ public class CommentaireController {
 		return ResponseEntity.status(HttpStatus.OK).body(listeCommentaire);
 	}
 
-	@PutMapping("/addcommentaire/{idPhoto}/{commentaireBox}")
+	@GetMapping("/getcommentairesbyphotoid/{photoId}")
+	public ResponseEntity<?> getCommentairesByPhotoId(@PathVariable Long photoId) {
+		List<Commentaire> listeCommentaire = photoRepo.findById(photoId).get().getCommentaires();
+		try {
+			listeCommentaire = (List<Commentaire>) commentaireRepo.findAll();
+		} catch (Exception e) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+		}
+		return ResponseEntity.status(HttpStatus.OK).body(listeCommentaire);
+	}
+	
+
+	@PutMapping("/addcommentaire/{idPhoto}/{photoPseudo}/{commentaireBox}")
 	public ResponseEntity<?> addCommentaire(@PathVariable Long idPhoto,
+											@PathVariable String photoPseudo,
 											@PathVariable String commentaireBox) {
 
 		try {
 			
 			return ResponseEntity.status(HttpStatus.OK)
-					.body(this.commentaireService.addCommentaire(idPhoto, commentaireBox));
+					.body(this.commentaireService.addCommentaire(idPhoto, photoPseudo, commentaireBox));
 		} catch (Exception e) {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
 		}
 
 	}
 
-	@DeleteMapping("/deletecommentaire")
-	public ResponseEntity<?> delCommentaire(@Valid long delId) {
-
-		this.commentaireRepo.deleteById(delId);
-
-		List<Commentaire> comList = null;
-		try {
-			comList = (List<Commentaire>) commentaireRepo.findAll();
+	@DeleteMapping("/deletecommentaire/{delId}/{pseudo}")
+	public ResponseEntity<?> delCommentaire(@PathVariable Long delId,
+											@PathVariable String pseudo) {
+		
+	try{
+	    			
+		if (this.commentaireRepo.findById(delId).get().getPhotographe().getPseudo().equals( pseudo) || (this.photographeRepo.findByPseudo(pseudo).getPseudo().equals( pseudo)))
+		
+			this.commentaireRepo.deleteById(delId);
+		
 		} catch (Exception e) {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
 		}
-		return ResponseEntity.status(HttpStatus.OK).body(comList);
+		return ResponseEntity.status(HttpStatus.OK).body(null);
 
-	}
+		}
+
 
 }
